Modified Hash to be 64bit

+++ src/Core/Reflection/ClassDescriptor.cpp
@@ -1,7 +1,7 @@
 #include "Core/Reflection/ClassDescriptor.h"
 
 
-ClassDescriptor::ClassDescriptor(std::string_view className, uint32_t typeID, size_t size)
+ClassDescriptor::ClassDescriptor(std::string_view className, uint64_t typeID, size_t size)
     : ClassName(className)
     , TypeID(typeID)
     , Size(size)
